+++ package.json
@@ -19,6 +19,7 @@
     "gl-matrix": "^3.4.3",
     "lodash": "^4.17.21",
     "next": "12.1.0",
+    "phosphor-react": "^1.4.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-redux": "^7.2.6",
